refactor: スクリプトを複数のモジュールに分割
