+++ modules/logo.txt
@@ -1,0 +1,7 @@
+ ____        _____           _
+|  _ \ _   _|  ___|   _  ___| | __
+| |_) | | | | |_ | | | |/ __| |/ /
+|  __/| |_| |  _|| |_| | (__|   <
+|_|    \__, |_|   \__,_|\___|_|\_\\
+       |___/
+                                    by satoki, xryuseix
